# Split alter_n into dedicated add and subtract logic.