@marshallpierce 2017-04-12
See also https://github.com/HdrHistogram/HdrHistogram/pull/123/ for equivalent subtraction logic change in the Java impl.