@marshallpierce 2017-04-12
The first set of benchmarks is before this change, the second is after. This change is a performance loss (but correctness win) for subtraction, but on the bright side, recording is faster.