@jonhoo 2017-04-12
What makes you say that recording is faster? Seems slower to me in all four `record_` tests?