@marshallpierce 2017-04-12
:thinking: I'm now questioning my sanity, but to me it looks like the units are ns/iteration, not iterations/sec, so smaller is faster.